+++ railties/test/generators/app_generator_test.rb
@@ -477,21 +477,13 @@ class AppGeneratorTest < Rails::Generators::TestCase
   def test_config_database_is_added_by_default
     run_generator
     assert_file "config/database.yml", /sqlite3/
-    if defined?(JRUBY_VERSION)
-      assert_gem "activerecord-jdbcsqlite3-adapter"
-    else
-      assert_gem "sqlite3", '">= 1.4"'
-    end
+    assert_gem "sqlite3", '">= 1.4"'
   end
 
   def test_config_mysql_database
     run_generator([destination_root, "-d", "mysql"])
     assert_file "config/database.yml", /mysql/
-    if defined?(JRUBY_VERSION)
-      assert_gem "activerecord-jdbcmysql-adapter"
-    else
-      assert_gem "mysql2", '"~> 0.5"'
-    end
+    assert_gem "mysql2", '"~> 0.5"'
   end
 
   def test_config_database_app_name_with_period
@@ -502,44 +494,7 @@ class AppGeneratorTest < Rails::Generators::TestCase
   def test_config_postgresql_database
     run_generator([destination_root, "-d", "postgresql"])
     assert_file "config/database.yml", /postgresql/
-    if defined?(JRUBY_VERSION)
-      assert_gem "activerecord-jdbcpostgresql-adapter"
-    else
-      assert_gem "pg", '"~> 1.1"'
-    end
-  end
-
-  def test_config_jdbcmysql_database
-    run_generator([destination_root, "-d", "jdbcmysql"])
-    assert_file "config/database.yml", /mysql/
-    assert_gem "activerecord-jdbcmysql-adapter"
-  end
-
-  def test_config_jdbcsqlite3_database
-    run_generator([destination_root, "-d", "jdbcsqlite3"])
-    assert_file "config/database.yml", /sqlite3/
-    assert_gem "activerecord-jdbcsqlite3-adapter"
-  end
-
-  def test_config_jdbcpostgresql_database
-    run_generator([destination_root, "-d", "jdbcpostgresql"])
-    assert_file "config/database.yml", /postgresql/
-    assert_gem "activerecord-jdbcpostgresql-adapter"
-  end
-
-  def test_config_jdbc_database
-    run_generator([destination_root, "-d", "jdbc"])
-    assert_file "config/database.yml", /jdbc/
-    assert_file "config/database.yml", /mssql/
-    assert_gem "activerecord-jdbc-adapter"
-  end
-
-  if defined?(JRUBY_VERSION)
-    def test_config_jdbc_database_when_no_option_given
-      run_generator
-      assert_file "config/database.yml", /sqlite3/
-      assert_gem "activerecord-jdbcsqlite3-adapter"
-    end
+    assert_gem "pg", '"~> 1.1"'
   end
 
   def test_generator_defaults_to_puma_version