+++ railties/CHANGELOG.md
@@ -1,2 +1,7 @@
+*   Remove support for `oracle`, `sqlserver` and JRuby specific database adapters from the
+    `rails new` and `rails db:system:change` commands. The supported options are `sqlite3`,
+    `mysql`, `postgresql` and `trilogy`.
+
+    *Andrew Novoselac*
 
 Please check [7-2-stable](https://github.com/rails/rails/blob/7-2-stable/railties/CHANGELOG.md) for previous changes.